Unified the parallel read functions
