+++ mexSrc/parallelreadzarr.cpp
@@ -1,202 +1,11 @@
 #include <omp.h>
 #include <fstream>
+#include "parallelreadzarrwrapper.h"
 #include "blosc2.h"
 #include "mex.h"
 #include "zarr.h"
 #include "helperfunctions.h"
 #include "zlib.h"
-
-void parallelReadZarrMex(const zarr &Zarr, void* zarrArr,
-                         const std::vector<uint64_t> &startCoords, 
-                         const std::vector<uint64_t> &endCoords,
-                         const std::vector<uint64_t> &readShape, const uint64_t bits)
-{
-    const uint64_t bytes = (bits/8);
-    
-    int32_t numWorkers = omp_get_max_threads();
-
-    blosc_init();
-    blosc_set_nthreads(numWorkers);
-    /*
-    if(numWorkers>Zarr.get_numChunks()){
-        blosc_set_nthreads(std::ceil(((double)numWorkers)/((double)Zarr.get_numChunks())));
-        numWorkers = Zarr.get_numChunks();
-    }
-    else {
-        blosc_set_nthreads(numWorkers);
-    }
-    */
-    
-    const int32_t batchSize = (Zarr.get_numChunks()-1)/numWorkers+1;
-    const uint64_t s = Zarr.get_chunks(0)*Zarr.get_chunks(1)*Zarr.get_chunks(2);
-    const uint64_t sB = s*bytes;
-
-    int err = 0;
-    std::string errString;
-
-    #pragma omp parallel for
-    for(int32_t w = 0; w < numWorkers; w++){
-        void* bufferDest = mallocDynamic(s,bits);
-        void* buffer = mallocDynamic(s,bits);
-        int64_t dsize = -1;
-        int uncErr = 0;
-        for(int64_t f = w*batchSize; f < (w+1)*batchSize; f++){
-            if(f>=Zarr.get_numChunks() || err) break;
-            const std::vector<uint64_t> cAV = Zarr.get_chunkAxisVals(Zarr.get_chunkNames(f));
-            const std::string subfolderName = Zarr.get_subfoldersString(cAV);
-
-            const std::string fileName(Zarr.get_fileName()+"/"+subfolderName+"/"+Zarr.get_chunkNames(f));
-
-            // If we cannot open the file then set to all zeros
-            // Can make this better by checking the errno
-            std::ifstream file(fileName, std::ios::binary);
-            if(!file.is_open()){
-                memset(bufferDest,0,sB);
-            }
-            else{
-                file.seekg(0, std::ios::end);
-                const std::streamsize filelen = file.tellg();
-                file.seekg(0, std::ios::beg);
-                file.read(reinterpret_cast<char*>(buffer), filelen);
-                file.close();
-
-                // Decompress
-                if(Zarr.get_cname() != "gzip"){
-                    dsize = blosc2_decompress(buffer, filelen, bufferDest, sB);
-                }
-                else{
-                    dsize = sB;
-                    z_stream stream;
-                    stream.zalloc = Z_NULL;
-                    stream.zfree = Z_NULL;
-                    stream.opaque = Z_NULL;
-                    stream.avail_in = (uInt)filelen;
-                    stream.avail_out = (uInt)dsize;
-                    while(stream.avail_in > 0){
-    
-                        dsize = sB;
-    
-                        stream.next_in = (uint8_t*)buffer+(filelen-stream.avail_in);
-                        stream.next_out = (uint8_t*)bufferDest+(sB-stream.avail_out);
-    
-                        uncErr = inflateInit2(&stream, 32);
-                        if(uncErr){
-                        #pragma omp critical
-                        {
-                        err = 1;
-                        errString = "Decompression error. Error code: "+
-                                     std::to_string(uncErr)+" ChunkName: "+
-                                     Zarr.get_fileName()+"/"+subfolderName+"/"+
-                                     Zarr.get_chunkNames(f)+"\n";
-                        }
-                        break;
-                        }
-        
-                        uncErr = inflate(&stream, Z_NO_FLUSH);
-        
-                        if(uncErr != Z_STREAM_END){
-                        #pragma omp critical
-                        {
-                        err = 1;
-                        errString = "Decompression error. Error code: "+
-                                     std::to_string(uncErr)+" ChunkName: "+
-                                     Zarr.get_fileName()+"/"+subfolderName+"/"+
-                                     Zarr.get_chunkNames(f)+"\n";
-                        }
-                        break;
-                        }
-                    }
-                    if(inflateEnd(&stream)){
-                        #pragma omp critical
-                        {
-                        err = 1;
-                        errString = "Decompression error. Error code: "+
-                                     std::to_string(uncErr)+" ChunkName: "+
-                                     Zarr.get_fileName()+"/"+subfolderName+"/"+
-                                     Zarr.get_chunkNames(f)+"\n";
-                        }
-                        break;
-                    }
-                }
-                
-                
-                if(dsize < 0){
-                    #pragma omp critical
-                    {
-                    err = 1;
-                    errString = "Decompression error. Error code: "+
-                                     std::to_string(uncErr)+" ChunkName: "+
-                                     Zarr.get_fileName()+"/"+subfolderName+"/"+
-                                     Zarr.get_chunkNames(f)+"\n";
-                    }
-                    break;
-                }
-            }
-            
-            if(Zarr.get_order() == "F"){
-                for(int64_t z = cAV[2]*Zarr.get_chunks(2); z < (cAV[2]+1)*Zarr.get_chunks(2); z++){
-                    if(z>=endCoords[2]) break;
-                    else if(z<startCoords[2]) continue;
-                    for(int64_t y = cAV[1]*Zarr.get_chunks(1); y < (cAV[1]+1)*Zarr.get_chunks(1); y++){
-                        if(y>=endCoords[1]) break;
-                        else if(y<startCoords[1]) continue;
-                        if(((cAV[0]*Zarr.get_chunks(0)) < startCoords[0] && ((cAV[0]+1)*Zarr.get_chunks(0)) > startCoords[0]) || (cAV[0]+1)*Zarr.get_chunks(0)>endCoords[0]){
-                            if(((cAV[0]*Zarr.get_chunks(0)) < startCoords[0] && ((cAV[0]+1)*Zarr.get_chunks(0)) > startCoords[0]) && (cAV[0]+1)*Zarr.get_chunks(0)>endCoords[0]){
-                                memcpy((uint8_t*)zarrArr+((((cAV[0]*Zarr.get_chunks(0))-startCoords[0]+(startCoords[0]%Zarr.get_chunks(0)))+((y-startCoords[1])*readShape[0])+((z-startCoords[2])*readShape[0]*readShape[1]))*bytes),(uint8_t*)bufferDest+(((startCoords[0]%Zarr.get_chunks(0))+((y%Zarr.get_chunks(1))*Zarr.get_chunks(0))+((z%Zarr.get_chunks(2))*Zarr.get_chunks(0)*Zarr.get_chunks(1)))*bytes),((endCoords[0]%Zarr.get_chunks(0))-(startCoords[0]%Zarr.get_chunks(0)))*bytes);
-                            }
-                            else if((cAV[0]+1)*Zarr.get_chunks(0)>endCoords[0]){
-                                memcpy((uint8_t*)zarrArr+((((cAV[0]*Zarr.get_chunks(0))-startCoords[0])+((y-startCoords[1])*readShape[0])+((z-startCoords[2])*readShape[0]*readShape[1]))*bytes),(uint8_t*)bufferDest+((((y%Zarr.get_chunks(1))*Zarr.get_chunks(0))+((z%Zarr.get_chunks(2))*Zarr.get_chunks(0)*Zarr.get_chunks(1)))*bytes),(endCoords[0]%Zarr.get_chunks(0))*bytes);
-                            }
-                            else if((cAV[0]*Zarr.get_chunks(0)) < startCoords[0] && ((cAV[0]+1)*Zarr.get_chunks(0)) > startCoords[0]){
-                                memcpy((uint8_t*)zarrArr+((((cAV[0]*Zarr.get_chunks(0)-startCoords[0]+(startCoords[0]%Zarr.get_chunks(0))))+((y-startCoords[1])*readShape[0])+((z-startCoords[2])*readShape[0]*readShape[1]))*bytes),(uint8_t*)bufferDest+(((startCoords[0]%Zarr.get_chunks(0))+((y%Zarr.get_chunks(1))*Zarr.get_chunks(0))+((z%Zarr.get_chunks(2))*Zarr.get_chunks(0)*Zarr.get_chunks(1)))*bytes),(Zarr.get_chunks(0)-(startCoords[0]%Zarr.get_chunks(0)))*bytes);
-                            }
-                        }
-                        else{
-                            memcpy((uint8_t*)zarrArr+((((cAV[0]*Zarr.get_chunks(0))-startCoords[0])+((y-startCoords[1])*readShape[0])+((z-startCoords[2])*readShape[0]*readShape[1]))*bytes),(uint8_t*)bufferDest+((((y%Zarr.get_chunks(1))*Zarr.get_chunks(0))+((z%Zarr.get_chunks(2))*Zarr.get_chunks(0)*Zarr.get_chunks(1)))*bytes),Zarr.get_chunks(0)*bytes);
-                        }
-                    }
-                }
-                
-            }
-            else if (Zarr.get_order() == "C"){
-                for(int64_t x = cAV[0]*Zarr.get_chunks(0); x < (cAV[0]+1)*Zarr.get_chunks(0); x++){
-                    if(x>=endCoords[0]) break;
-                    else if(x<startCoords[0]) continue;
-                    for(int64_t y = cAV[1]*Zarr.get_chunks(1); y < (cAV[1]+1)*Zarr.get_chunks(1); y++){
-                        if(y>=endCoords[1]) break;
-                        else if(y<startCoords[1]) continue;
-                        for(int64_t z = cAV[2]*Zarr.get_chunks(2); z < (cAV[2]+1)*Zarr.get_chunks(2); z++){
-                            if(z>=endCoords[2]) break;
-                            else if(z<startCoords[2]) continue;
-                            switch(bytes){
-                                case 1:
-                                    ((uint8_t*)zarrArr)[((x-startCoords[0])+((y-startCoords[1])*readShape[0])+((z-startCoords[2])*readShape[0]*readShape[1]))] = ((uint8_t*)bufferDest)[((z%Zarr.get_chunks(2))+((y%Zarr.get_chunks(1))*Zarr.get_chunks(2))+((x%Zarr.get_chunks(0))*Zarr.get_chunks(2)*Zarr.get_chunks(1)))];
-                                    break;
-                                case 2:
-                                    ((uint16_t*)zarrArr)[((x-startCoords[0])+((y-startCoords[1])*readShape[0])+((z-startCoords[2])*readShape[0]*readShape[1]))] = ((uint16_t*)bufferDest)[((z%Zarr.get_chunks(2))+((y%Zarr.get_chunks(1))*Zarr.get_chunks(2))+((x%Zarr.get_chunks(0))*Zarr.get_chunks(2)*Zarr.get_chunks(1)))];
-                                    break;
-                                case 4:
-                                    ((float*)zarrArr)[((x-startCoords[0])+((y-startCoords[1])*readShape[0])+((z-startCoords[2])*readShape[0]*readShape[1]))] = ((float*)bufferDest)[((z%Zarr.get_chunks(2))+((y%Zarr.get_chunks(1))*Zarr.get_chunks(2))+((x%Zarr.get_chunks(0))*Zarr.get_chunks(2)*Zarr.get_chunks(1)))];
-                                    break;
-                                case 8:
-                                    ((double*)zarrArr)[((x-startCoords[0])+((y-startCoords[1])*readShape[0])+((z-startCoords[2])*readShape[0]*readShape[1]))] = ((double*)bufferDest)[((z%Zarr.get_chunks(2))+((y%Zarr.get_chunks(1))*Zarr.get_chunks(2))+((x%Zarr.get_chunks(0))*Zarr.get_chunks(2)*Zarr.get_chunks(1)))];
-                                    break;
-                            }
-                            
-                        }
-                    }
-                }
-                
-            }
-            
-        }
-        free(bufferDest);
-        free(buffer);
-    }
-    blosc_destroy();
-
-    if(err) mexErrMsgIdAndTxt("zarr:threadError",errString.c_str());
-}
 
 // TODO: FIX MEMORY LEAKS
 void mexFunction(int nlhs, mxArray *plhs[],
